pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-tool-factory"
version = "0.1.0"
description = "Generate universal MCP servers from natural language, OpenAPI specs, or Python functions"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Hesham Salama", email = "info@heshamsalama.dev" }
]
keywords = ["mcp", "ai", "tools", "agents", "llm", "openai", "claude", "langchain"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mcp>=1.0.0",
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# LLM Providers (install at least one)
anthropic = [
    "anthropic>=0.40.0",
]
claude-code = [
    "claude-agent-sdk>=0.1.0",  # Claude Agent SDK with OAuth token support
]
openai = [
    "openai>=1.50.0",
]
google = [
    "google-generativeai>=0.8.0",
]
# Development
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
openapi = [
    "openapi-spec-validator>=0.7.0",
]
# All providers
all-providers = [
    "mcp-tool-factory[anthropic,claude-code,openai,google]",
]
all = [
    "mcp-tool-factory[anthropic,claude-code,openai,google,dev,openapi]",
]

[project.scripts]
mcp-factory = "tool_factory.cli:cli"

[project.urls]
Homepage = "https://github.com/HeshamFS/mcp-tool-factory"
Documentation = "https://github.com/HeshamFS/mcp-tool-factory#readme"
Repository = "https://github.com/HeshamFS/mcp-tool-factory"

[tool.hatch.build.targets.wheel]
packages = ["src/tool_factory"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=tool_factory --cov-report=term-missing --cov-report=html --cov-fail-under=85"

[tool.coverage.run]
source = ["src/tool_factory"]
branch = true
omit = [
    "*/__pycache__/*",
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
